move wasmd encoding cfg to fn
