+++ cosmwasmclient/config/query_config_test.go
